gettrip/requirements.txt
asgiref==3.8.0
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
pillow==10.2.0
sqlparse==0.4.4
